best move is already found when only one move hit a fail high alphabeta search
It is a search algorithm improvement based on the fact that for playing
one have to find the best move, not it's evaluation.
When a single move hit a fail high at the root of alphabeta search,
it is necessarely the best one. In this case increasing the aspiration
window is costly by doing useless subsequent search just to look for
the exact evaluation of the best move.

An additional modification was to setup the initial upper bound of aspiration
search at previousbestmove.previousScore+1, such as when none of the other
moves reach this value, all search leading to an improvement above the
previous score is not done.